ajuste put couchdb

+++ src/couchdb/couchdb.service.ts
@@ -7,26 +7,64 @@ export class CouchdbService {
   private db: Nano.DocumentScope<any>;
 
   constructor() {
-
     this.nano = Nano('http://admin:senha@localhost:5984');
     this.db = this.nano.use('api_loja'); 
   }
 
   async createDocument(doc: any) {
-    return await this.db.insert(doc);
+    try {
+      return await this.db.insert(doc);
+    } catch (error: any) {
+      console.error('Erro ao criar documento:', error.message);
+      throw new Error('Não foi possível criar o documento.');
+    }
   }
 
   async getDocument(id: string) {
-    return await this.db.get(id);
+    try {
+      return await this.db.get(id);
+    } catch (error: any) {
+      if (error.statusCode === 404) {
+        console.warn(`Documento com ID '${id}' não encontrado.`);
+        return null;
+      }
+      console.error('Erro ao buscar documento:', error.message);
+      throw new Error('Não foi possível buscar o documento.');
+    }
   }
 
   async updateDocument(id: string, doc: any) {
-    const existingDoc = await this.getDocument(id);
-    return await this.db.insert({ ...existingDoc, ...doc });
+    try {
+      const existingDoc = await this.getDocument(id);
+      if (!existingDoc) {
+        throw new Error(`Documento com ID '${id}' não encontrado.`);
+      }
+
+      const updatedDoc = {
+        ...existingDoc,
+        ...doc,
+        _id: id,
+        _rev: existingDoc._rev,
+      };
+
+      return await this.db.insert(updatedDoc);
+    } catch (error: any) {
+      console.error('Erro ao atualizar documento:', error.message);
+      throw new Error('Não foi possível atualizar o documento.');
+    }
   }
 
   async deleteDocument(id: string) {
-    const existingDoc = await this.getDocument(id);
-    return await this.db.destroy(id, existingDoc._rev);
+    try {
+      const existingDoc = await this.getDocument(id);
+      if (!existingDoc) {
+        throw new Error(`Documento com ID '${id}' não encontrado.`);
+      }
+
+      return await this.db.destroy(id, existingDoc._rev);
+    } catch (error: any) {
+      console.error('Erro ao deletar documento:', error.message);
+      throw new Error('Não foi possível deletar o documento.');
+    }
   }
 }
